resolved warnings

+++ src/GameObject/ParticleManager.cpp
@@ -126,7 +126,7 @@ void ParticleManager::render() {
 void ParticleManager::createExplosionParticle(float x, float y) {
 	const float randSize = rand() % static_cast<int>(PARTICLE_MAX_WIDTH - PARTICLE_MIN_WIDTH) + PARTICLE_MIN_WIDTH * 10;
 	const float randAngle = rand() % 360 * Math::m_PI / 180.f;
-	const float randDistance = rand() % maxPosOffset;
+	const float randDistance = static_cast<float>(rand() % maxPosOffset);
 
 	const ParticleManager::Particle newParticle = ParticleManager::Particle{
 		true,
@@ -146,7 +146,7 @@ void ParticleManager::createExplosionParticle(float x, float y) {
 void ParticleManager::createFireworkParticle(float x, float y, float explosionRadius) {
 	const float randSize = rand() % static_cast<int>(PARTICLE_MAX_WIDTH - PARTICLE_MIN_WIDTH) + PARTICLE_MIN_WIDTH * 20;
 	const float randAngle = rand() % 360 * Math::m_PI / 180.f;
-	const float randRadius = rand() % static_cast<int>(explosionRadius);
+	const float randRadius = static_cast<float>(rand() % static_cast<int>(explosionRadius));
 
 	// Generate random color
 	const float red = static_cast<float>(rand()) / RAND_MAX;
